Seperate main and mirrorApp.
